pyproject.toml
[tool.poetry]
name = "factuality"
version = "1.0.13"
description = ""
authors = ["Felix Mönckemeyer <felix.moenckemeyer@rwth-aachen.de>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
gpt-json = "^0.4.2"
newspaper3k = "^0.2.8"
pytablewriter = "^1.2.0"
structlog = "^24.1.0"
rich = "^13.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
factuality = "factuality.__main__:main"
